# RentalManagement/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "RentalManagement API",
    "description": "An ASP.NET Core Web API for managing rentals",
    "version": "v1"
  },
  "paths": {
    "/api/Places": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Gets a place by ID from the database",
        "responses": {
          "200": {
            "description": "The places were found.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Places"
        ],
        "summary": "Creates a new place",
        "requestBody": {
          "description": "The place to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaceDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaceDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaceDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              }
            }
          },
          "201": {
            "description": "The place was created."
          },
          "400": {
            "description": "The place is invalid."
          }
        }
      }
    },
    "/api/Places/{placeId}": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Gets a place by ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place to get",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The place was found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              }
            }
          },
          "404": {
            "description": "The place was not found."
          }
        }
      },
      "put": {
        "tags": [
          "Places"
        ],
        "summary": "Updates a place by ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated information of the place",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaceDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaceDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaceDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The place was updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDTO"
                }
              }
            }
          },
          "400": {
            "description": "The place is invalid."
          },
          "404": {
            "description": "The place was not found."
          }
        }
      },
      "delete": {
        "tags": [
          "Places"
        ],
        "summary": "Deletes a place by ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "Place to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "204": {
            "description": "The place was deleted."
          },
          "404": {
            "description": "The place was not found."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreatePlaceDTO": {
        "type": "object",
        "properties": {
          "roomsCount": {
            "type": "integer",
            "format": "int32",
            "example": 3
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "example": 100
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "1234 Main St, Springfield, IL 62701"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Beautiful place with a view of the park"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 1000
          }
        },
        "additionalProperties": false
      },
...
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
      "PlaceDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "roomsCount": {
            "type": "integer",
            "format": "int32",
            "example": 3
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "example": 100
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "1234 Main St, Springfield, IL 62701"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Beautiful place with a view of the park"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 100
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePlaceDTO": {
        "type": "object",
        "properties": {
          "roomsCount": {
            "type": "integer",
            "format": "int32",
            "example": 3
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "example": 100
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "1234 Main St, Springfield, IL 62701"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Beautiful place with a view of the park"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 100
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}